+++ inc/template-tags.php
@@ -420,6 +420,7 @@ if ( ! function_exists( 'loose_comment' ) ) :
 		$custom_css .= '.menu-logo {float:left;}';
 		$custom_css .= '.navbar-navigation ul, .nav-social {display:block;}';
 		$custom_css .= '.loose-featured-slider, .loose-featured-slider .featured-image, .loose-featured-slider .no-featured-image {height:' . absint( get_theme_mod( 'home_page_slider_height', 500 ) ) . 'px;}';
+		$custom_css .= '.entry-content a {color:' . esc_attr( get_theme_mod( 'content_link_color', '#000' ) ) . ';}';
 
 		wp_add_inline_style( 'loose-style', $custom_css );
 		}
@@ -655,7 +656,7 @@ if ( ! function_exists( 'loose_comment' ) ) :
 			} elseif ( ! is_single() && has_post_format( 'quote' ) ) {
 				$subject = get_the_content();
 				$subject = apply_filters( 'the_content', $subject );
-				$regex = '/<cite>.*<\/cite>/';
+				$regex = '/<cite>.*<\/cite>/s';
 				preg_match( $regex, $subject, $matches );
 				if ( $matches && $matches[0] ) {
 					$match = $matches[0];